feat: drop support for macOS Sierra and below
Drop support for macOS Sierra and below as we are no longer testing this OS. Drop a line on stderr on each CLI command and inform the user that this OS is not supported.

Change the public API (getSystemWarnings) to return objects with information about the severity of the warnings. This way the consumer can decide how to handle the warnings.

Fix an issue when getting macOS version from System Profile - the code was not working when version has two numbers (10.14 for example).

Add new method in logger that allows printing to stderr. It should be used instead of `console.error`, so we can make the code testable.